Adding async-await for generateId function
When providing an Id from an in-memory database, accessing the database over the network and getting the id may cause to skip `id` assignment. 

In this case id will be `undefined`. Because of this reason engine.io fall into an endless loop.
To overcome this issue async-await combination can be used. 

Also async-await solution is not a breaking change.